requirements.txt
torch>=2.1
transformers>=4.42
datasets>=2.20
evaluate>=0.4.2
pyyaml>=6.0
loguru>=0.7.2
tqdm>=4.66
tensorboard>=2.16
